Move `Edges` to its own module
