+++ packages/graph/hash_graph/lib/graph/src/shared/subgraph/mod.rs
@@ -1,12 +1,10 @@
 use std::{
-    collections::{
-        hash_map::{IntoIter, RawEntryMut},
-        HashMap, HashSet,
-    },
+    collections::{HashMap, HashSet},
     fmt::{Debug, Formatter},
 };
 
 use depths::GraphResolveDepths;
+use edges::Edges;
 use serde::{Deserialize, Serialize};
 use type_system::{DataType, EntityType, PropertyType};
 use utoipa::{openapi, ToSchema};
@@ -19,6 +17,7 @@ use crate::{
 };
 
 mod depths;
+mod edges;
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize)]
 #[serde(rename_all = "camelCase")]
@@ -62,26 +61,6 @@ impl ToSchema for Vertex {
 
         builder.into()
     }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, ToSchema)]
-#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-pub enum EdgeKind {
-    /// An entity has a link
-    HasLink,
-    /// A link has an entity as its destination
-    HasDestination,
-    /// A link or entity has a link type or entity type as its type, respectively
-    HasType,
-    /// A type can reference another type
-    References,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct OutwardEdge {
-    pub edge_kind: EdgeKind,
-    pub destination: GraphElementId,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, ToSchema)]
@@ -143,52 +122,5 @@ where
             .field("filter", &self.filter)
             .field("graph_resolve_depths", &self.graph_resolve_depths)
             .finish()
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Default, Serialize)]
-pub struct Edges(HashMap<GraphElementId, HashSet<OutwardEdge>>);
-
-impl Edges {
-    #[must_use]
-    pub fn new() -> Self {
-        Self(HashMap::new())
-    }
-
-    pub fn insert(&mut self, identifier: GraphElementId, edge: OutwardEdge) -> bool {
-        match self.0.raw_entry_mut().from_key(&identifier) {
-            RawEntryMut::Vacant(entry) => {
-                entry.insert(identifier, HashSet::from([edge]));
-                true
-            }
-            RawEntryMut::Occupied(entry) => {
-                let set = entry.into_mut();
-                set.insert(edge)
-            }
-        }
-    }
-}
-
-// Necessary because utoipa can't handle HashSet
-impl ToSchema for Edges {
-    fn schema() -> openapi::Schema {
-        openapi::ObjectBuilder::new()
-            .additional_properties(Some(openapi::schema::Array::new(OutwardEdge::schema())))
-            .into()
-    }
-}
-
-impl IntoIterator for Edges {
-    type IntoIter = IntoIter<GraphElementId, HashSet<OutwardEdge>>;
-    type Item = (GraphElementId, HashSet<OutwardEdge>);
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.0.into_iter()
-    }
-}
-
-impl Extend<(GraphElementId, HashSet<OutwardEdge>)> for Edges {
-    fn extend<T: IntoIterator<Item = (GraphElementId, HashSet<OutwardEdge>)>>(&mut self, other: T) {
-        self.0.extend(other);
     }
 }
