+++ packages/graph/hash_graph/lib/graph/src/store/postgres/ontology/property_type.rs
@@ -10,13 +10,19 @@ use crate::{
     identifier::ontology::OntologyTypeEditionId,
     ontology::{OntologyElementMetadata, PropertyTypeWithMetadata},
     provenance::{CreatedById, OwnedById, UpdatedById},
-    shared::{identifier::GraphElementId, subgraph::depths::GraphResolveDepths},
+    shared::{
+        identifier::GraphElementId,
+        subgraph::{
+            depths::GraphResolveDepths,
+            edges::{EdgeKind, OutwardEdge},
+        },
+    },
     store::{
         crud::Read,
         postgres::{context::PostgresContext, DependencyContext, DependencyContextRef},
         AsClient, InsertionError, PostgresStore, PropertyTypeStore, QueryError, UpdateError,
     },
-    subgraph::{EdgeKind, OutwardEdge, StructuralQuery, Subgraph},
+    subgraph::{StructuralQuery, Subgraph},
 };
 
 impl<C: AsClient> PostgresStore<C> {
